webpack modifications
Split webpack into modules basically for production and dev. Created a production module for webpack which is very basic as webpack defaults most of what is needed. Added some production script to package.json.

ToDo: Look into modules for uglifying the bundle, lazy loading, chunking etc... for production. Still a work in progress but it is very useable right now.